// src/app/features/feed/interfaces/IAllPostsResponse.ts
export interface User {
	_id: string;
	name: string;
	username: string;
	photo: string;
}

export interface CommentCreator {
	_id: string;
	name: string;
	username: string;
	photo: string;
}

export interface TopComment {
	_id: string;
	content: string;
	commentCreator: CommentCreator;
	post: string;
	parentComment?: any;
	likes: any[];
	createdAt: string;
}

export interface Post {
	_id: string;
	body: string;
	privacy: string;
	user: User;
	image: string;
	sharedPost?: any;
	likes: any[];
	createdAt: string;
	commentsCount: number;
	topComment: TopComment;
	sharesCount: number;
	likesCount: number;
	isShare: boolean;
	id: string;
	bookmarked: boolean;
}

export interface Data {
	posts: Post[];
}

export interface Pagination {
	currentPage: number;
	numberOfPages: number;
	limit: number;
	nextPage: number;
	total: number;
}

export interface Meta {
	pagination: Pagination;
}

export interface IAllPostsResponse {
	success: boolean;
	message: string;
	data: Data;
	meta: Meta;
}